Remove searchresult.json file, move operator to operator.py, usic async requests, use assetType enum
