+++ src/application/application.py
@@ -10,7 +10,7 @@ class Application(object):
         """Create an Application object."""
 
     def background(self) -> bool:
-        """Return True when blender is running without a user interface (started withk -b).
+        """Return True when blender is running without a user interface (started with -b).
 
         Returns:
             bool: background
